feat: add simple endpoint to bootstrap sparkjava
Sparkjava is a lightweight framework for creating HTTP based
applications. This is the library I intend to use for implementing the
API.

Currently exposing a single route `/ordermanagement/orders` that returns
a dummy text response. Note the lack of version number in the URI.
Google "Rich Hickey versioning" for a much more articulate explanation
of why than I could give.