test: Create JUnit Rule for sparkjava lifecycle management
In the unit tests for the HTTP API definitions we want to spin up a
working sparkjava server configured with the endpoint under test.

This server should utilize a random port to avoid conflicting with other
sparkjava instances either of the application running, parallel tests or
sequential unit tests that run too fast (shutdown of the server is
currently asynchronous).

This JUnit Rule implements this logic in a manner that is trivial to
 reuse between tests/projects with minimal boilerplate code.

An example of usage will come in a followup commit.