refactor: make repository return Option when looking up Order by id
The `Table.getItem` API returns null if the item could not be found.

Returning an Option allows the calling code to handle this somewhat gracefully, not that I claim the current Service code is particularly graceful.

Ideally this would use the `Try` construct from javaslang but I'm not going to go that far with changes tonight.